requirements.txt
label-studio-sdk>=0.0.29
requests>=2.31.0
python-dotenv==1.0.0
tenacity>=8.2.3
tabulate>=0.9.0
pathlib


